Kafka/requirements.txt
boto3==1.39.1
botocore==1.39.1
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
jmespath==1.0.1
kafka-python==2.2.15
kafka-python-ng==2.2.3
numpy==2.3.1
pandas==2.3.0
py4j==0.10.9.9
pyarrow==20.0.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
wldhx.yadisk-direct==0.0.6
